# doc/api.yaml
openapi: 3.0.0
info:
  title: WASATEXT
  description: API for WASATEXT messaging platform
  version: 1.0.0
...
security:
  - WASATextAuth: []
servers:
  - url: http://0.0.0.0:3000
    description: Local Server
paths:
...
  /users/{userId}/conversations/{conversationId}/messages/{messageId}/comments:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/conversationId"
      - $ref: "#/components/parameters/MessageId"
    get:
      summary: Get all comments of a message
      description: From a message, the user can see all comments on this.
      operationId: getComments
      tags:
        - "Comment"

      responses:
        "200":
          description: Returns the comments
          content:
            application/json:
              schema:
                description: List of comments
                properties:
                  comments:
                    description: List schema
                    type: array
                    minItems: 0
                    maxItems: 2048
                    items:
                      $ref: "#/components/schemas/Comment"
              example:
                comments:
                  - commentId: "100446"
                    emoji: "❤️"
                    time: "2025-10-31T15:42:00Z"
                  - commentId: "100447"
                    emoji: "👍"
                    time: "2025-10-31T16:10:00Z"

        "500":
          $ref: "#/components/responses/InternalError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "400":
          $ref: "#/components/responses/BadRequest"


    post:
      summary: Comment Message
      description: From a message, the user can react to it by adding an emoji reaction.
      operationId: commentMessage
      tags:
        - "Comment"

      requestBody:
        content:
          application/json:
            schema:
              description: Comment schema
              properties:
                emoji:
                  type: string
                  description: Unicode of emoji used in the comment.
                  pattern: "^[\u0000-\uFFFF]+$"
                  minLength: 0
                  maxLength: 1024
                  example: ☠️
        required: true
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
              example:
                commentId: "100448"
                emoji: "🔥"
                time: "2025-10-31T17:20:00Z"
        "500":
          $ref: "#/components/responses/InternalError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /users/{userId}/groups/{groupId}/info/name:
    put:
      summary: Set Group name
      description: The user can set a personalized group name.
      operationId: setGroupName
      tags:
        - "Group"
      parameters:
        - $ref: "#/components/parameters/userId"
        - $ref: "#/components/parameters/GroupId"
      requestBody:
        content:
          application/json:
            schema:
              description: Schema for updating the group's name
              properties:
                name:
                  description: Group's name
                  type: string
                  minLength: 1
                  maxLength: 255
                  pattern: "^[a-zA-Z0-9_ -]+$"
                  example: capodanno 2027
      responses:
        "201":
          description: Name updated.
          content:
            application/json:
              schema:
                description: Updated group name response schema
                properties:
                  name:
                    description: Group's name
                    type: string
                    minLength: 1
                    maxLength: 255
                    pattern: "^[a-zA-Z0-9_ -]+$"
                    example: capodanno_2027
                  time:
                    description: Timestamp
                    type: string
                    format: date-time
                    example: 2025-10-31T15:42:00Z
                    minLength: 20
                    maxLength: 20
                    pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}Z$"
              example:
                name: "capodanno_2027"
                time: "2025-10-31T15:42:00Z"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /users/{userId}/groups/{groupId}:
    get:
      summary: Show the conversation
      description: Returns the details of a single conversation identified by conversationId from the user's conversations.
      operationId: getGroup
      tags:
        - "Conversations"
      parameters:
        - $ref: "#/components/parameters/userId"
        - $ref: "#/components/parameters/GroupId"
      responses:
        "200":
          description: Returns the chat as a list of messages.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
              example:
                groupId: "288344"
                name: "capodanno2026"
                upload:
                  url: "https://www.stoccolmaaroma.it/wp-content/uploads/2013/07/logo-sapienza.jpg"
                  mime: "image/jpg"
                  width: 1024
                  height: 720
                participants:
                  - userId: "10012"
                    username: "Simone"
                  - userId: "77772"
                    username: "Alessio"


        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  schemas:
...
    UserId:
      description: Operation User's ID
      type: string
      pattern: "^[0-9]+$"
      minLength: 1
      maxLength: 9999
      example: "11300"
      readOnly: true
    Username:
      description: User's username
      type: string
      minLength: 1
      maxLength: 255
      pattern: "^[a-zA-Z0-9_ -]+$"
      example: Simone
    User:
      description: User schema
      type: object
      properties:
        userId:
          $ref: "#/components/schemas/UserId"
        username:
          $ref: "#/components/schemas/Username"
      required:
        - userId
        - username
    Group:
      type: object
      description: group schema
      properties:
        groupId:
          $ref: "#/components/schemas/GroupId"
        name:
          description: Group name
          type: string
          minLength: 0
          maxLength: 255
          pattern: "^[a-zA-Z0-9_ -]+$"
          example: capodanno2026
        upload:
          $ref: "#/components/schemas/Upload"
        participants:
          description: list of participants of a group
          type: array
          minItems: 2
          maxItems: 256
          items:
            $ref: "#/components/schemas/User"
          example:
            - userId: "11300"
              username: Simone
            - userId: "11301"
              username: Luca
    ConversationId:
      description: Unique ID assigned to each conversation resource in the system.
      pattern: "^[0-9]+$"
      minLength: 1
      maxLength: 9999
      example: "110056"
      readOnly: true
      type: string
    CommentId:
      description: Unique ID assigned to each comment resource in the system.
      pattern: "^[0-9]+$"
      minLength: 1
      maxLength: 9999
      example: "100446"
      readOnly: true
      type: string
    MessageId:
      description: Unique ID assigned to each message resource in the system.
      pattern: "^[0-9]+$"
      minLength: 1
      maxLength: 9999
      example: "2635444"
      readOnly: true
      type: string
    GroupId:
      description: Unique ID assigned to each group resource in the system.
      type: string
      pattern: "^[0-9]+$"
      minLength: 1
      maxLength: 9999
      example: "2773354"
      readOnly: true
    Comment:
      type: object
      description: Comment schema
      properties:
        commentId:
          $ref: "#/components/schemas/CommentId"
        emoji:
          type: string
          description: Unicode of emoji used in the comment.
          pattern: "^[\u0000-\uFFFF]+$"
          example: ☠️
        time:
          description: Timestamp
          type: string
          minLength: 20
          maxLength: 20
          pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}Z$"
          format: date-time
          example: 2025-10-31T15:42:00Z
...
    Upload:
      description: upload's schema for responses or view
      type: object
      properties:
        url:
          type: string
          minLength: 1
          maxLength: 255
          pattern: "^[a-zA-Z]+://.+$"
          format: url
          description: Photo's URL to access the image
          example: https://www.stoccolmaaroma.it/wp-content/uploads/2013/07/logo-sapienza.jpg
        mime:
          description: mime represents the type of the Photo(image/png...)
          type: string
          minLength: 0
          maxLength: 255
          pattern: '^[a-zA-Z]+/[a-zA-Z.]+$'
          example: image/jpg
        width:
          description: width of the image
          type: integer
          example: 1024
        height:
          description: height of the image
          type: integer
          example: 720
...
    Error:
      description: Info of an error
      type: object
      properties:
        code:
          description: code of the error
          type: integer
          example: 401
        message:
          description: text of the error
          type: string
          minLength: 0
          maxLength: 2048
          example: Unauthorized Action, request invalid or missing.
      required:
        - code
        - message
  parameters:
    userId:
      name: userId
      in: path
      required: true
      description: User Identifier
      schema:
        $ref: "#/components/schemas/UserId"
    conversationId:
      name: conversationId
      in: path
      required: true
      description: Chat Identifier
      schema:
        $ref: "#/components/schemas/ConversationId"
    MessageId:
      name: messageId
      in: path
      required: true
      description: Message Identifier
      schema:
        $ref: "#/components/schemas/MessageId"
...
    GroupId:
      name: groupId
      in: path
      required: true
      description: Group Identifier
      schema:
        $ref: "#/components/schemas/GroupId"
  securitySchemes:
    WASATextAuth:
      type: http
      description: Protect operation session token.
      scheme: bearer
      bearerFormat: SessionToken
  responses:
    UnauthorizedError:
      description: Unauthorized Action, request invalid or missing.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 401
            message: "Unauthorized Action, request invalid or missing."
    InternalError:
      description: An error occurred, please retry.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 500
            message: "An error occurred, please retry."
    NotFoundError:
      description: Resource not found. Check the parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 404
            message: "Resource not found. Check the parameters."
    BadRequest:
      description: Bad Request. Invalid parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 400
            message: "Bad Request. Invalid parameters."
...